+++ src/main/java/weare4saken/config/Config.java
@@ -1,4 +1,4 @@
-package weare4saken;
+package weare4saken.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
