+++ src/main/java/weare4saken/model/Car.java
@@ -1,28 +1,30 @@
-package weare4saken;
+package weare4saken.model;
 
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
+import weare4saken.service.Transport;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 
 @Component
-@Scope("prototype")
-public class Moto implements Transport {
+@Scope("singleton")
+public class Car implements Transport {
 
     @Override
     public void go() {
-        System.out.println("Мотоцикл едет");
+        System.out.println("Автомобиль едет");
     }
 
     @PostConstruct
     public void init() {
-        System.out.println("Class Moto: init method");
+        System.out.println("Class Car: init method");
     }
 
     @PreDestroy
     public void destroy() {
-        System.out.println("Class Moto: destroy method");
+        System.out.println("Class Car: destroy method");
     }
 
 }
+
