+++ src/main/java/weare4saken/model/Moto.java
@@ -1,29 +1,29 @@
-package weare4saken;
+package weare4saken.model;
 
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
+import weare4saken.service.Transport;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 
 @Component
-@Scope("singleton")
-public class Car implements Transport{
+@Scope("prototype")
+public class Moto implements Transport {
 
     @Override
     public void go() {
-        System.out.println("Автомобиль едет");
+        System.out.println("Мотоцикл едет");
     }
 
     @PostConstruct
     public void init() {
-        System.out.println("Class Car: init method");
+        System.out.println("Class Moto: init method");
     }
 
     @PreDestroy
     public void destroy() {
-        System.out.println("Class Car: destroy method");
+        System.out.println("Class Moto: destroy method");
     }
 
 }
-
