+++ src/main/resources/applicationContext.xml
@@ -8,11 +8,11 @@
     http://www.springframework.org/schema/context/spring-context.xsd">
 
     <bean id="customCar"
-          class="weare4saken.Car">
+          class="weare4saken.model.Car">
     </bean>
 
     <bean id="customPerson"
-          class="weare4saken.Person">
+          class="weare4saken.model.Person">
         <property name="car" ref="customCar"/>
     </bean>
